Added process.env() call
A simple implementation that returns an array of strings with the key and value separated by an '='. Saves an external exec to `env` or `set`. A smarter
implementation might return the value of an env var with a single argument, or null if it doesn't exist, and setenv a variable if 2 args are passed, or remove the
env var if the 2nd arg is null.

Also a small change to Makefile.osx to add linking to -lsqlite3